fix: three correctness bugs — cell-list pair-drop (#4699), f64 dtype crash, PBC tie-rounding (#4232) (#14)
* fix: three correctness bugs — cell-list pair-drop (#4699), f64 dtype crash, PBC tie-rounding (#4232)

Fix 1a (backlog #4699): jax_md cell-list silently drops in-cutoff pairs
- Changed disable_cell_list default from False to True in:
  - src/prolix/physics/neighbor_list.py:160
  - src/prolix/physics/nonbonded_config.py:26
- Root cause: jax_md.partition.neighbor_list's cell-list path silently drops
  real in-cutoff atom pairs at protein+solvent density (~134 atoms/cell
  on 1vii system) with NO overflow flag. Brute-force O(N²) is safer.

Fix 1b (debt 832): hardcoded float32 cast crashes under jax_enable_x64=True
- Removed hardcoded .astype(jnp.float32) casts from:
  - src/prolix/physics/explicit_corrections.py:77, 116
  - src/prolix/physics/flash_nonbonded.py:165, 168-170, 274, 277-279, 307, 351, 399
- Root cause: JAX's grad accumulation requires cotangents to exactly match
  primal dtype. Hardcoded float32 intermediates clashed with float64 primals
  under x64, causing "lax.add requires arguments to have the same dtypes".
- Fix: Use ambient/input dtype (positions.dtype, epsilon.dtype, etc.) instead.

Fix 1c (backlog #4232): PBC minimum-image tie-rounding non-determinism
- Replaced jnp.round (banker's rounding) with round-half-away-from-zero
  in all minimum-image computations:
  - src/prolix/physics/pbc.py:50 (minimum_image_distance)
  - src/prolix/physics/flash_explicit.py:188, 299 (chunked kernels)
- Root cause: When pair separation equals box/2, jnp.round's round-half-to-even
  breaks ties by nearest-even parity, leading to non-deterministic wrap direction
  depending on floating-point rounding errors.
- Fix: Implement round-half-away-from-zero (standard MD convention) via
  sign(x)*floor(abs(x)+0.5), ensuring deterministic PBC wrapping.

Verification:
- test_pbc.py: added test_minimum_image_distance_at_box_half_boundary (passes)
- test_tail_corrections.py: added f64 autodiff tests (pass)
- test_1vii_nl_nonbonded_parity.py: all 3 tests pass (NL path unaffected)


* fix(test): strengthen PBC tie-rounding regression test (#4232)

The added test_minimum_image_distance_at_box_half_boundary only checked
scalar distance at an exact box/2 separation, which is invariant under
both round-half-to-even and round-half-away-from-zero (both conventions
wrap to a mirror-symmetric dr of the same magnitude here) — so it would
have passed identically before the fix too. Add a direct test of
_round_half_away_from_zero against jnp.round's actual tie behavior,
which is the only way to discriminate the two conventions.


---------

+++ src/prolix/physics/explicit_corrections.py
@@ -74,7 +74,8 @@ def lj_dispersion_tail_energy(
     box = jnp.asarray(box)
     sigma = jnp.asarray(sigma)
     epsilon = jnp.asarray(epsilon)
-    atom_mask = jnp.asarray(atom_mask).astype(jnp.float32)
+    # debt 832: dtype follows epsilon (the ambient float precision) rather than hardcoding float32.
+    atom_mask = jnp.asarray(atom_mask).astype(epsilon.dtype)
     volume = box[0] * box[1] * box[2] if box.ndim == 1 else jnp.linalg.det(box)
 
     sum_c6 = jnp.sum(jnp.sqrt(epsilon) * (sigma**3) * atom_mask)
@@ -113,7 +114,8 @@ def lj_dispersion_tail_impulsive_pressure(
     box = jnp.asarray(box)
     sigma = jnp.asarray(sigma)
     epsilon = jnp.asarray(epsilon)
-    atom_mask = jnp.asarray(atom_mask).astype(jnp.float32)
+    # debt 832: dtype follows epsilon (the ambient float precision) rather than hardcoding float32.
+    atom_mask = jnp.asarray(atom_mask).astype(epsilon.dtype)
     volume = box[0] * box[1] * box[2] if box.ndim == 1 else jnp.linalg.det(box)
 
     sum_c6 = jnp.sum(jnp.sqrt(epsilon) * (sigma**3) * atom_mask)
